RENAME: sample -> slatekit.status
