+++ src/lib/kotlin/slatekit-status/src/jvmMain/kotlin/slatekit/status/SampleJvm.kt
@@ -1,4 +1,4 @@
-package sample
+package slatekit.status
 
 actual class Sample {
     actual fun checkMe() = 42
