+++ src/lib/kotlin/slatekit-status/src/jvmTest/kotlin/slatekit/status/SampleTestsJVM.kt
@@ -1,4 +1,4 @@
-package sample
+package slatekit.status
 
 import kotlin.test.Test
 import kotlin.test.assertTrue
