+++ src/lib/kotlin/slatekit-status/src/macosMain/kotlin/slatekit/status/SampleMacos.kt
@@ -1,4 +1,4 @@
-package sample
+package slatekit.status
 
 actual class Sample {
     actual fun checkMe() = 7
